Rename new module to avoid possible conflict with Python standard library
Fix #227
