+++ fastetl/hooks/db_to_db_hook.py
@@ -10,7 +10,7 @@ from typing import Dict, Optional
 from airflow.hooks.base import BaseHook
 
 from fastetl.custom_functions.fast_etl import copy_db_to_db, sync_db_2_db
-from fastetl.types import DBSource
+from fastetl.data_types import DBSource
 
 
 class DbToDbHook(BaseHook):
